fix: types and build
