+++ @remirror/preset-table/src/table-preset.ts
@@ -11,19 +11,13 @@ import {
  * The table is packaged up as preset for simpler
  */
 export class TablePreset extends Preset {
-  static documentReady = false;
-  get name() {
-    return 'table' as const;
-  }
-
-  init() {
-    if (TablePreset.documentReady) {
-      return;
-    }
-
+  static enable() {
     document.execCommand('enableObjectResizing', false, 'false');
     document.execCommand('enableInlineTableEditing', false, 'false');
-    TablePreset.documentReady = true;
+  }
+
+  get name() {
+    return 'table' as const;
   }
 
   createExtensions() {
